+++ src/HomeView/CardContainer/CardContainer.scss
@@ -1,4 +1,4 @@
-@import '../index.scss';
+@import '../../index.scss';
 
 .card-container {
   @include row-wrap
