+++ src/HomeView/CardContainer/PlantCard/PlantCard.scss
